refactor: 重构账单单位为原来的元修改为分

+++ src/main/java/com/zhuche/server/model/Transaction.java
@@ -25,7 +25,7 @@ public class Transaction extends BaseEntity{
 
     private Double amount;
 
-    private Double balance; // 余额
+    private int balance; // 余额，单位: 分
 
     private String title;
